# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Server pets",
        "title": "Swagger Example API",
        "contact": {},
        "license": {},
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/stock/all/list": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stock"
                ],
                "summary": "获取所有列表",
                "parameters": [
                    {
                        "description": "获取所有列表",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AllDetailListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/stock/monitor/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "StockMonitor"
                ],
                "summary": "添加监控",
                "parameters": [
                    {
                        "description": "添加监控",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AddMonitorReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"添加监控成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/stock/monitor/del": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "StockMonitor"
                ],
                "summary": "删除监控",
                "parameters": [
                    {
                        "description": "删除监控",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DelMonitorReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"删除监控成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/stock/monitor/list": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "StockMonitor"
                ],
                "summary": "获取监控列表",
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/stock/monitor/one": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "StockMonitor"
                ],
                "summary": "获取监控",
                "parameters": [
                    {
                        "description": "获取监控",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.MonitorOneReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/stock/monitor/update": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "StockMonitor"
                ],
                "summary": "更新监控",
                "parameters": [
                    {
                        "description": "更新监控",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateMonitorReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"更新监控成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.AddMonitorReq": {
            "type": "object",
            "properties": {
                "monitor_high": {
                    "type": "number"
                },
                "monitor_low": {
                    "type": "number"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "request.AllDetailListReq": {
            "type": "object",
            "properties": {
                "page_num": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
...
        "request.DelMonitorReq": {
            "type": "object",
            "properties": {
                "symbol": {
                    "type": "string"
                }
            }
        },
...
        "request.MonitorOneReq": {
            "type": "object",
            "properties": {
                "symbol": {
                    "type": "string"
                }
            }
        },
...
        "request.UpdateMonitorReq": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "monitor_high": {
                    "type": "number"
                },
                "monitor_low": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-token",
            "in": "header"
        }
    }
}